[kahi_ranking_udea_works] working in support for no dois papers and in the stucture to the plugin
