+++ Kahi_ranking_udea_works/kahi_ranking_udea_works/parser.py
@@ -1,0 +1,50 @@
+from kahi_impactu_utils.Utils import doi_processor, lang_poll, split_names
+from time import time
+from pandas import isna
+
+
+def parse_ranking_udea(reg, affiliation, empty_work):
+    entry = empty_work.copy()
+    entry["updated"] = [{"source": "ranking_udea", "time": int(time())}]
+    lang = lang_poll(reg["titulo"])
+    entry["titles"].append(
+        {"title": reg["titulo"], "lang": lang, "source": "ranking_udea"})
+    if reg["DOI"]:
+        if not isna(reg["DOI"]):
+            doi = doi_processor(reg["DOI"])
+            if doi:
+                entry["external_ids"].append(
+                    {"source": "doi", "id": doi})
+    if reg["issn"]:
+        if isinstance(reg["issn"], str):
+            for issn in reg["issn"].strip().split():
+                if "-" not in issn:
+                    continue
+                issn = issn.strip()
+                entry["source"] = {"name": reg["nombre rev o premio"],
+                                   "external_ids": [{"source": "issn", "id": issn}]}
+    if not entry["source"]:
+        entry["source"] = {
+            "name": reg["nombre rev o premio"], "external_ids": []}
+    entry["year_published"] = int(reg["año realiz"])
+    name = split_names(reg["nombre"])
+    aff = {
+        "id": affiliation["_id"],
+        "name": affiliation["names"][0]["name"],
+        "types": affiliation["types"]
+    }
+    for affname in affiliation["names"]:
+        if affname["lang"] == "es":
+            aff["name"] = affname["name"]
+            break
+        elif affname["lang"] == "en":
+            aff["name"] = affname["name"]
+        elif affname["source"] == "ror":
+            aff["name"] = affname["name"]
+    entry["authors"].append({
+        "external_ids": [{"source": "Cédula de Ciudadanía", "id": reg["cedula"]}],
+        "full_name": name["full_name"],
+        "types": [],
+        "affiliations": [aff]
+    })
+    return entry
